modif float to integer (ty tim)

+++ db/migrate/20220523104341_create_reservations.rb
@@ -5,7 +5,7 @@ class CreateReservations < ActiveRecord::Migration[7.0]
       t.references :offer, null: false, foreign_key: true
       t.date :start_date
       t.date :end_date
-      t.float :total_price
+      t.integer :total_price
 
       t.timestamps
     end
